requirements.txt
vastai-sdk>=0.3.0
boto3>=1.34.0
